Shale/1.1.0/makefile pedantic++

+++ sample/simple-accelerometer/src/accel.h
@@ -6,6 +6,7 @@ extern "C" {
 #endif
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #define ACCEL_SUCCESS 0
 #define ACCEL_PARAM_ERROR -1
@@ -60,8 +61,9 @@ struct accelState;
  */
 typedef int (*accel_callback)(struct accelState *state, int gesture_id, int offset_found, bool *reset_gesture);
 
+// TODO: define this as a partial-type instead of exposing some fields.
 typedef struct accelState {
-    int dimensions;
+    uint32_t dimensions;
 
     accel_callback callback;
     struct internalAccelState *state;
@@ -84,8 +86,8 @@ typedef struct accelState {
  *                     gestures must be before the callback is called.
  * @return             ACCEL_SUCCESS if successful, an error code otherwise.
  */
-int accel_generate_state(accel_state **state, int dimensions, int window_size, accel_callback callback,
-                         const int threshold);
+int accel_generate_state(accel_state **state, uint32_t dimensions, uint16_t window_size, accel_callback callback,
+                         const uint32_t threshold);
 
 /**
  * Destroys the state object at the pointer pointed to by the state pointer.
@@ -96,30 +98,30 @@ int accel_generate_state(accel_state **state, int dimensions, int window_size, a
 int accel_destroy_state(accel_state **state);
 
 /**
- * Starts recording a accel gesture
+ * Starts recording an accel gesture
  * @param  state   A pointer to a non-NULL state variable that holds recording
  *                 metadata.
  * @param  gesture Non-NULL pointer that will be populated with the gesture id.
  * @return         ACCEL_SUCCESS if successful, an error code otherwise.
  */
-int accel_start_record_gesture(accel_state *state, int *gesture);
+int accel_start_record_gesture(accel_state *state, uint16_t *gesture);
 
 /**
- * Ends recording a accel gesture
+ * Ends recording an accel gesture
  * @param state      A pointer to a non-NULL state variable that holds recording
  *                   metadata.
  * @param gesture_id Value that corresponds to a gesture currently being
  *                   recorded.
  * @return           ACCEL_SUCCESS if successful, an error code otherwise.
  */
-int accel_end_record_gesture(accel_state *state, int gesture_id);
+int accel_end_record_gesture(accel_state *state, uint16_t gesture_id);
 
 /**
- * Updates the state variable's current state based on the accel data array
- * passed in.
+ * Updates the state variable's current state based on the raw accelerometer
+ * data array passed in.
  * @param  state      A pointer to a non-NULL state variable that holds
  *                    recording metadata.
- * @param  accel_data An with accelerometer data.
+ * @param  accel_data A state->dimensions array with accelerometer data.
  * @return            ACCEL_SUCCESS if successful, an error code otherwise.
  */
 int accel_process_timer_tick(accel_state *state, int *accel_data);
@@ -135,7 +137,7 @@ int accel_process_timer_tick(accel_state *state, int *accel_data);
  *                    distance corresponding to the returned gesture.
  * @return            ACCEL_SUCCESS if successful, an error code otherwise.
  */
-int accel_find_most_likely_gesture(accel_state *state, int *gesture_id, int *distance);
+int accel_find_most_likely_gesture(accel_state *state, uint16_t *gesture_id, int *distance);
 
 /**
  * For a given state and recorded gesture, resets the gesture's offset state
