+++ sample/simple-accelerometer/src/moving_avg_ticker.h
@@ -5,6 +5,8 @@
 extern "C" {
 #endif
 
+#include <stdint.h>
+
 #include "accel.h"
 
 #define MOVING_AVG_SUCCESS ACCEL_SUCCESS
@@ -16,20 +18,20 @@ typedef struct moving_avg_values {
     // Circular buffer
     int *wbuf;
     int wbuf_end;
-    int wbuf_len;
+    uint16_t wbuf_len;
 
     int subtotal;
     int subtotal_size;
     int max_subtotal_size;
 } moving_avg_values;
 
-int allocate_moving_avg(int num_wbuf, int subtotal_sizes, moving_avg_values **allocated);
+int allocate_moving_avg(uint16_t num_wbuf, int subtotal_sizes, moving_avg_values **allocated);
 
 int reset_moving_avg(moving_avg_values *reset);
 
 int append_to_moving_avg(moving_avg_values *value, int appended, bool *isAtEnd);
 
-int get_latest_frame_moving_avg(moving_avg_values *value, int *frame);
+int get_latest_frame_moving_avg(moving_avg_values *value, int32_t *frame);
 
 int free_moving_avg(moving_avg_values **value);
